src/requirements.txt
PyContracts